paho-mqtt-c: Update to the real 1.3.2, and correct the license
In commit b0345e3f (paho-mqtt-c: upgrade 1.3.1 -> 1.3.2), paho-mqtt-c
was supposedly updated to 1.3.2. However, the SRCREV did not match the
v1.3.2 tag.

With the update to 1.3.2, the license was changed from EPL-1.0 | EDL-1.0
to EPL-2.0 | EDL-1.0. This was updated in commit b0345e3f for
LIC_FILES_CHKSUM, but not in LICENSE.

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>